+++ components/persistor/lib/remote-doc/remote-doc-types/RemoteDocClient.ts
@@ -1,0 +1,5 @@
+export interface RemoteDocClient {
+    uploadDocument(base64: string, key: string, bucket: string);
+    downloadDocument(key: string, bucket: string);
+    deleteDocument(key: string, bucket: string);
+}