+++ components/persistor/lib/types/PersistorTransaction.ts
@@ -1,0 +1,9 @@
+export type PersistorTransaction = {
+    id: number,
+    dirtyObjects: object,
+    savedObjects: object,
+    touchObjects: object,
+    deletedObjects: object,
+    remoteObjects: Set<string> // identifiers for objects not stored directly in our db
+    deleteQueries: {}
+};