+++ components/persistor/lib/types/RemoteDocConnectionOptions.ts
@@ -1,0 +1,4 @@
+export type RemoteDocConnectionOptions = {
+    bucketName: string,
+    environment: string
+}