+++ components/persistor/lib/types/index.ts
@@ -1,0 +1,2 @@
+export { PersistorTransaction } from './PersistorTransaction';
+export { RemoteDocConnectionOptions } from './RemoteDocConnectionOptions'